refactor: promote compactList display to separate component
