+++ components/skill-edit/SkillEditItem.tsx
@@ -1,9 +1,8 @@
-import { IconButton, Grid, makeStyles } from '@material-ui/core';
 import { Dispatch, SetStateAction } from 'react';
 import { InputField } from '..';
 import { ComponentWithT, Skill } from '../../lib/types';
 import { LevelSlider } from '../level-slider/LevelSlider';
-import { DeleteOutline } from '@material-ui/icons';
+import { CompactListItem } from '../compact-list-item/CompactListItem';
 
 interface SkillEditItemProps extends ComponentWithT {
   skills: Record<string, Skill>;
@@ -12,15 +11,6 @@ interface SkillEditItemProps extends ComponentWithT {
   setSkill: Dispatch<SetStateAction<Record<string, Skill>>>;
 }
 
-const useStyles = makeStyles({
-  deleteContainer: {
-    paddingTop: '2rem',
-  },
-  button: {
-    marginRight: '-2rem',
-  },
-});
-
 export const SkillEditItem = ({
   skills,
   id,
@@ -28,8 +18,7 @@ export const SkillEditItem = ({
   t,
   type = 'skill',
 }: SkillEditItemProps): React.ReactElement => {
-  const classes = useStyles();
-  const handleClick = (): void => {
+  const handleDelete = (): void => {
     setSkill((prev) => {
       const newSkills = { ...prev };
       delete newSkills[id];
@@ -47,35 +36,22 @@ export const SkillEditItem = ({
       />
     ) : null;
   return (
-    <Grid container alignItems="flex-start">
-      <Grid xs={10} item>
-        <InputField
-          label={t(`labels.${type}.name`)}
-          propName={[id, 'name']}
-          value={skills[id].name}
-          setValue={
-            setSkill as Dispatch<SetStateAction<Record<string, unknown>>>
-          }
-        />
-        <LevelSlider
-          t={t}
-          type={type}
-          label={t(`labels.${type}.level`)}
-          propName={[id, 'level']}
-          setValue={
-            setSkill as Dispatch<SetStateAction<Record<string, unknown>>>
-          }
-          input={skills[id].level}
-        />
-      </Grid>
-      <Grid item xs={2}>
-        <Grid container justify="flex-end" className={classes.deleteContainer}>
-          <IconButton className={classes.button} onClick={handleClick}>
-            <DeleteOutline />
-          </IconButton>
-        </Grid>
-      </Grid>
+    <CompactListItem handleDelete={handleDelete}>
+      <InputField
+        label={t(`labels.${type}.name`)}
+        propName={[id, 'name']}
+        value={skills[id].name}
+        setValue={setSkill as Dispatch<SetStateAction<Record<string, unknown>>>}
+      />
+      <LevelSlider
+        t={t}
+        type={type}
+        label={t(`labels.${type}.level`)}
+        propName={[id, 'level']}
+        setValue={setSkill as Dispatch<SetStateAction<Record<string, unknown>>>}
+        input={skills[id].level}
+      />
       {description}
-    </Grid>
+    </CompactListItem>
   );
 };
